# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Terramaps API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/maps/{map_id}/invites": {
      "get": {
        "tags": [
          "Invites"
        ],
        "summary": "List Map Invites",
        "description": "List pending invites for a map. Any map member.",
        "operationId": "list_map_invites_maps__map_id__invites_get",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MapInvite"
                  },
                  "title": "Response List Map Invites Maps  Map Id  Invites Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Invites"
        ],
        "summary": "Create Invite",
        "description": "Invite a user by email to a map. Any map member.",
        "operationId": "create_invite_maps__map_id__invites_post",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapInvite"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/maps/{map_id}/invites/{invite_id}": {
      "delete": {
        "tags": [
          "Invites"
        ],
        "summary": "Revoke Invite",
        "description": "Revoke a pending invite. Any map member.",
        "operationId": "revoke_invite_maps__map_id__invites__invite_id__delete",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "invite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invite Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/invites/{invite_id}/accept": {
      "post": {
        "tags": [
          "Invites"
        ],
        "summary": "Accept Invite",
        "description": "Accept a pending invite. Grants MEMBER access and marks the invite accepted.",
        "operationId": "accept_invite_me_invites__invite_id__accept_post",
        "parameters": [
          {
            "name": "invite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invite Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/invites/{invite_id}/decline": {
      "post": {
        "tags": [
          "Invites"
        ],
        "summary": "Decline Invite",
        "description": "Decline a pending invite.",
        "operationId": "decline_invite_me_invites__invite_id__decline_post",
        "parameters": [
          {
            "name": "invite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invite Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateInviteDTO": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "CreateInviteDTO",
        "description": "Request body for creating a map invite."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MapInvite": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "map_id": {
            "type": "string",
            "title": "Map Id"
          },
          "invited_email": {
            "type": "string",
            "title": "Invited Email"
          },
          "invited_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invited By Name"
          },
          "invited_by_email": {
            "type": "string",
            "title": "Invited By Email"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "declined"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "map_id",
          "invited_email",
          "invited_by_name",
          "invited_by_email",
          "status"
        ],
        "title": "MapInvite",
        "description": "Pending invite as seen by the map owner."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}